+++ CHANGES.rst
@@ -18,6 +18,10 @@ Changes
 - Enable coveralls.io for coverage measurement and run doctests on all
   supported Python versions.
 
+- Fix `issue 9
+  <https://github.com/zopefoundation/zope.security/issues/9>`_:
+  iteration of ``itertools.groupby`` objects is now allowed by default.
+
 4.1.1 (2017-05-17)
 ------------------